add logic for recoverIntepreters
